requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
install==1.3.5
python-telegram-bot==13.10
pytz==2024.1
six==1.16.0
tornado==6.4
tzlocal==5.2
